Adding REST controllers skeletons and user soft delete refactor

+++ backend/src/main/java/filmly/dto/contentrating/ContentRatingResponseDto.java
@@ -6,6 +6,6 @@ public record ContentRatingResponseDto(
         Long contentRatingId,
         Long contentId,
         Long userId,
-        Double rating,
+        Float rating,
         LocalDateTime createdAt
 ) {}
